Large update commit - new features & fixing bugs
This commit introduces a large update to the project.

There is a new Input System handling the Inputs from the real-fixed-based driving simulator via "Vjoy"-Plugin. It handles the inputs from the Steering Wheel, Brakes and Gas-Pedals.

It also handles the input from the buttons on the driving-simulator.

The commit contains a new feature, allowing the 3D-Pass-Through-Mask-Car-Model in Unity to be aligned with the real-world model of the fixed-based driving simulator. This works by aligning both with the use of Varjo Markers and calculating the exact offset between the real world object and the game object.

It also allows for intuitive adjustment of the offset-positioning of the masks. This makes the adjustment process for new markers quick and easy.

There is a new feature making the transition between multiple markers smooth and noticeable to the user.

The 3D-Model of the car is updated and is now more precise. The car seat is also deleted.